third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-search-replace-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests the search replace functionality.


Running: testReplaceAll
--- Before replace ---
// var a1, a2, a3;
const  a1, a2, a3;

--- After replace ---
//= var a$$, b$&, a3 /* $0 $1 $2 $& $$ \0 \1 */;
/** @const */ var a$$, b$&, a3 /* $0 $1 $2 $& $$ \0 \1 */;
